requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.8.2
llvmlite==0.44.0
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.65.1
openai-whisper==20240930
packaging==24.2
pathspec==0.12.1
pi==0.1.2
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-telegram-bot==21.10
regex==2024.11.6
requests==2.32.3
sentry-sdk==2.29.1
sniffio==1.3.1
soundfile==0.13.1
sympy==1.13.1
tiktoken==0.9.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2025.8.22